move sim logic from motion.rs to simulation/mod.rs
Simulation logic was still in `motion.rs`, because I forgot to move it
during one of the major refactors.
