+++ src/simulation/motion.rs
@@ -1,30 +1,6 @@
-use super::collect_perfromance_data;
-use super::collisions;
-use super::gravity;
 use crate::particle::Particle;
-use crate::simulation::SimSettings;
-use crate::PHYSICS_UPDATE_HZ;
 
 use bevy::prelude::*;
-
-pub struct MotionPlugin;
-
-impl Plugin for MotionPlugin {
-    fn build(&self, app: &mut App) {
-        app.add_systems(
-            FixedUpdate,
-            (
-                gravity::calc_grav_accel,
-                collisions::calculate_collisions,
-                update_particle_positions,
-                collect_perfromance_data,
-            )
-                .chain()
-                .run_if(sim_not_paused),
-        )
-        .insert_resource(Time::<Fixed>::from_hz(PHYSICS_UPDATE_HZ));
-    }
-}
 
 #[derive(Component)]
 pub struct OldPosition(pub Transform);
@@ -32,12 +8,7 @@ pub struct OldPosition(pub Transform);
 #[derive(Component)]
 pub struct Acceleration(pub Vec2);
 
-/// returns true if the simulation is not paused
-fn sim_not_paused(settings: Res<SimSettings>) -> bool {
-    !settings.paused
-}
-
-fn update_particle_positions(
+pub fn update_particle_positions(
     mut query: Query<(&mut Transform, &mut OldPosition, &mut Acceleration), With<Particle>>,
     time: Res<Time>,
 ) {
